requirements.txt
feedparser
google-generativeai
requests
python-dotenv
